Added File menu event handler

+++ src/items_panel.py
@@ -4,6 +4,7 @@ import re
 import os
 from list_pop_menu import ListItemPopUp
 import list_pop_menu
+import settings
 
 class ItemPanel(object):
     """
@@ -19,16 +20,7 @@ class ItemPanel(object):
         self.items = []
         self.list_control = wx.ListCtrl(self.frame.content_splitter,
                                    style=wx.LC_REPORT|wx.LC_SINGLE_SEL)
-        self.list_control.InsertColumn(0, 'Title')
-        self.list_control.InsertColumn(1, 'User Name')
-        self.list_control.InsertColumn(2, 'Password')
-        self.list_control.InsertColumn(3, 'Notes')
-
-        self.list_control.SetColumnWidth(0, 100)
-        self.list_control.SetColumnWidth(1, 200)
-        self.list_control.SetColumnWidth(2, 150)
-        self.list_control.SetColumnWidth(3, 420)
-
+        self.set_columns()
         self.image_list = wx.ImageList(width=10, height=10)
         self.image_list.Add(wx.Bitmap(self.frame.icon_dir+self.icon_name,
                                       type=wx.BITMAP_TYPE_PNG)
@@ -41,13 +33,36 @@ class ItemPanel(object):
         
         # Register the application for required events
         self.register_events()
+
+    def set_columns(self):
+        """
+            Description: Sets the column name for list panel
+        """
+        self.list_control.InsertColumn(0, 'Title')
+        self.list_control.InsertColumn(1, 'User Name')
+        self.list_control.InsertColumn(2, 'Password')
+        
+        self.list_control.SetColumnWidth(0, 100)
+        self.list_control.SetColumnWidth(1, 200)
+        self.list_control.SetColumnWidth(2, 150)
         
     def register_events(self):
         """
             Description: Register the required events for this Window
         """
         self.list_control.Bind(wx.EVT_LIST_ITEM_RIGHT_CLICK, self.show_pop_menu)
+        self.list_control.Bind(wx.EVT_LIST_ITEM_SELECTED, self.show_details)
+        self.list_control.Bind(wx.EVT_LIST_ITEM_DESELECTED, self.show_details)
 
+    def show_details(self, event):
+        """
+            Description: Called on event when the user selects/deselects
+                            a list item in the list panel
+            input_param: event - Select/De-Select Event 
+            input_type: Event instance
+        """
+        self.frame.detail_panel.show_details(event)
+    
     def copy_item(self, event):
         """
             Description: Called on event when the user clicks
@@ -57,11 +72,11 @@ class ItemPanel(object):
         """
         item_index = self.list_control.GetFocusedItem()
         menu_id = event.GetId()
-        if wx.TheClipboard.Open():
+        if wx.TheClipboard.Open() and item_index != -1:
             wx.TheClipboard.Clear()
-            if menu_id == list_pop_menu.ITEM_COPY_USER:
+            if menu_id == settings.ITEM_COPY_USER:
                 wx.TheClipboard.SetData(wx.TextDataObject(self.list_control.GetItem(item_index, 1).GetText()))
-            elif menu_id == list_pop_menu.ITEM_COPY_PASS:
+            elif menu_id == settings.ITEM_COPY_PASS:
                 wx.TheClipboard.SetData(wx.TextDataObject(self.list_control.GetItem(item_index, 2).GetText()))
             else:
                 event.Skip()
@@ -75,7 +90,7 @@ class ItemPanel(object):
             input_type: Event instance
         """
         self.pop_up_menu.layout_pop_menu()
-
+    
     def delete_item(self, event):
         """
             Description: Called on event when the user click delete
@@ -86,9 +101,12 @@ class ItemPanel(object):
         folder_item = self.frame.folder_panel.folders_control.GetFocusedItem()
         folder_data = self.frame.folder_panel.folders_control.GetPyData(folder_item)
         item_index = self.list_control.GetFocusedItem()
-        index = self.list_control.GetItemData(item_index)
-        self.list_control.DeleteItem(item_index)
-        del folder_data['items'][index] 
+        if item_index != -1:
+            index = self.list_control.GetItemData(item_index)
+            self.list_control.DeleteItem(item_index)
+            del folder_data['items'][index]
+            self.frame.content_saved = False
+            self.frame.set_title(self.frame.file_name)
     
     def edit_item(self, event):
         """
@@ -100,12 +118,14 @@ class ItemPanel(object):
         folder_item = self.frame.folder_panel.folders_control.GetFocusedItem()
         folder_data = self.frame.folder_panel.folders_control.GetPyData(folder_item)
         item_index = self.list_control.GetFocusedItem()
-        index = self.list_control.GetItemData(item_index)
-        item_data = folder_data['items'][index]
-        # Todo: Edit Dialog to enter the details
-        self.list_control.GetItem(item_index, 1).SetText('poda')
-        
-
+        if item_index != -1:
+            index = self.list_control.GetItemData(item_index)
+            item_data = folder_data['items'][index]
+            # Todo: Edit Dialog to enter the details
+            self.list_control.GetItem(item_index, 1).SetText('poda')
+            self.frame.content_saved = False
+            self.frame.set_title(self.frame.file_name)
+    
     def display_items(self):
         """
             Description: Displays the list of credential items on the panel
@@ -117,8 +137,7 @@ class ItemPanel(object):
             self.list_control.InsertStringItem(row_id, item['title'])
             self.list_control.SetStringItem(row_id, 1, item['name'])
             self.list_control.SetStringItem(row_id, 2, item['password'])
-            self.list_control.SetStringItem(row_id, 3, item['notes'])
-            self.list_control.SetItemImage(row_id, item.get('item_icon_id', 0))
+            self.list_control.SetItemImage(row_id, int(item.get('item_icon_id', 0)))
             self.list_control.SetItemData(row_id, index)
             if (row_id % 2) == 0:
                 self.list_control.SetItemBackgroundColour(row_id, '#e6f1f5')
@@ -134,12 +153,15 @@ class ItemPanel(object):
         if new_item['folder'] in self.frame.folder_panel.folder_details.keys():
             self.frame.folder_panel.folder_details[new_item['folder']]['items'].append(new_item)
         else:
-            self.frame.folder_panel.folder_details[new_item['folder']] = {
-                            'items' : [new_item],
-                            'icon_id': 0,
-                            'path': new_item.get('folder', '/')
+            new_folder = {
+                'icon_id': 0,
+                'path': new_item.get('folder', '/'),
+                'items': []
             }
-
+            if new_item.get('title'):
+                new_folder['items'] = [new_item]
+            self.frame.folder_panel.folder_details[new_item['folder']] = new_folder
+            
     def set_column_value(self, new_item, line):
         """
             Description: Assign the column value into the new item dictionary
@@ -211,14 +233,17 @@ class ItemPanel(object):
             file_content = ''
             secret_key = 'RKLOCK'
             for folder, folder_details in self.frame.folder_panel.folder_details.items():
-                for item in folder_details.get('items', []):
-                    file_content = file_content + '[' + item['title'] + ']\r\n'
-                    file_content = file_content + 'User Name: ' + item['name'] + '\r\n'
-                    file_content = file_content + 'Password: ' + item['password'] + '\r\n'
-                    file_content = file_content + 'Notes: ' + item['notes'] + '\r\n'
-                    file_content = file_content + 'Group: ' + item['folder'] + '\r\n'
-                    file_content = file_content + 'Item Icon' + str(item.get('item_icon_id', 0)) + '\r\n'
-                    file_content = file_content + '\r\n'
+                if len( folder_details.get('items', [])) > 0:
+                    for item in folder_details.get('items', []):
+                        file_content = file_content + '[' + item['title'] + ']\r\n'
+                        file_content = file_content + 'User Name: ' + item['name'] + '\r\n'
+                        file_content = file_content + 'Password: ' + item['password'] + '\r\n'
+                        file_content = file_content + 'Notes: ' + item['notes'] + '\r\n'
+                        file_content = file_content + 'Group: ' + item['folder'] + '\r\n'
+                        file_content = file_content + 'Item Icon: ' + str(item.get('item_icon_id', 0)) + '\r\n'
+                        file_content = file_content + '\r\n'
+                else:
+                    file_content = file_content + 'Group: ' + folder + '\r\n\r\n'
 
             if len(file_content)%self.frame.block_size != 0:
                 padding = ''.zfill(
@@ -232,5 +257,5 @@ class ItemPanel(object):
                     total_contet = secret_key+file_content
                     pass_file.write(des_obj.encrypt(total_contet))
             event.Skip()
-        except:
-            event.Skip()
+        except Exception as e:
+            pass
